refactor: adjustments due to upstream's entry point changes
